function/f1-radio-collector/requirements.txt
functions-framework==3.10.1
google-cloud-bigquery==3.41.0
google-cloud-storage==3.10.1
google-cloud-logging==3.15.0
requests==2.33.1
python-dotenv==1.2.2
